+++ hazzy/modules/camview/camview.py
@@ -17,7 +17,6 @@
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
 
-import cv2
 import gtk
 import gobject
 import threading
@@ -32,11 +31,7 @@ import gettext
 
 _ = gettext.gettext
 
-# check if opencv3 is used, so we will have to change attribut naming
-from pkg_resources import parse_version
-
-OPCV3 = parse_version(cv2.__version__) >= parse_version('3')
-
+from video import VideoDev
 
 class CamView(gtk.VBox):
     """
@@ -104,72 +99,10 @@ class CamView(gtk.VBox):
         'exit': (gobject.SIGNAL_RUN_FIRST, gobject.TYPE_NONE, ()),
     }
 
-    def __init__(self, videodevice=0, frame_width=640, frame_height=480):
+    def __init__(self, video_device):
         super(CamView, self).__init__()
 
         self.__version__ = "0.1.1"
-
-        # set the selected camera as video device
-        self.videodevice = videodevice
-
-        # set the capture size
-        self.frame_width = frame_width
-        self.frame_height = frame_height
-
-        # set other default values or initialize them
-        self.color = (0, 0, 255)
-        self.radius = 150
-        self.radius_difference = 10
-        self.circles = 5
-        self.autosize = False
-        self.full_info = False
-        self.linewidth = 1
-        # self.frame = None
-        self.img_gtk = None
-        self.paused = False
-        self.thrd = None
-        self.initialized = False
-
-        self.img_width = self.frame_width
-        self.img_height = self.frame_height
-        self.img_ratio = float(self.img_width) / float(self.img_height)
-
-        self.colorseldlg = None
-        self.old_frames = 0
-
-        # set the correct camera as video device
-        self.cam = cv2.VideoCapture(self.videodevice)
-
-        # 0  = CAP_PROP_POS_MSEC        Current position of the video file in milliseconds.
-        # 1  = CAP_PROP_POS_FRAMES      0-based index of the frame to be decoded/captured next.
-        # 2  = CAP_PROP_POS_AVI_RATIO   Relative position of the video file
-        # 3  = CAP_PROP_FRAME_WIDTH     Width of the frames in the video stream.
-        # 4  = CAP_PROP_FRAME_HEIGHT    Height of the frames in the video stream.
-        # 5  = CAP_PROP_FPS             Frame rate.
-        # 6  = CAP_PROP_FOURCC          4-character code of codec.
-        # 7  = CAP_PROP_FRAME_COUNT     Number of frames in the video file.
-        # 8  = CAP_PROP_FORMAT          Format of the Mat objects returned by retrieve() .
-        # 9  = CAP_PROP_MODE            Backend-specific value indicating the current capture mode.
-        # 10 = CAP_PROP_BRIGHTNESS      Brightness of the image (only for cameras).
-        # 11 = CAP_PROP_CONTRAST        Contrast of the image (only for cameras).
-        # 12 = CAP_PROP_SATURATION      Saturation of the image (only for cameras).
-        # 13 = CAP_PROP_HUE             Hue of the image (only for cameras).
-        # 14 = CAP_PROP_GAIN            Gain of the image (only for cameras).
-        # 15 = CAP_PROP_EXPOSURE        Exposure (only for cameras).
-        # 16 = CAP_PROP_CONVERT_RGB     Boolean flags indicating whether images should be converted to RGB.
-        # 17 = CAP_PROP_WHITE_BALANCE   Currently unsupported
-        # 18 = CAP_PROP_RECTIFICATION   Rectification flag for stereo cameras
-        #                               (note: only supported by DC1394 v 2.x backend currently)
-        if OPCV3:
-            self.cam.set(cv2.CAP_PROP_FRAME_WIDTH, self.frame_width)
-            self.cam.set(cv2.CAP_PROP_FRAME_HEIGHT, self.frame_height)
-        else:
-            self.cam.set(cv2.cv.CV_CAP_PROP_FRAME_WIDTH, self.frame_width)
-            self.cam.set(cv2.cv.CV_CAP_PROP_FRAME_HEIGHT, self.frame_height)
-
-        self.cam_properties = CamProperties()
-        self.cam_properties.get_devices()
-        # self.cam_properties.get_resolution(self.videodevice)
 
         # make the main GUI
         self.image_box = gtk.EventBox()
@@ -234,8 +167,10 @@ class CamView(gtk.VBox):
         self.adj_videodevice = gtk.Adjustment(0, 0, len(self.cam_properties.devices) - 1, 1, 1, 0)
         self.adj_videodevice.connect("value_changed", self.adj_videodevice_value_changed)
         self.spn_videodevice = gtk.SpinButton(self.adj_videodevice, 0, 0)
+
         if len(self.cam_properties.devices) == 1:
             self.spn_videodevice.set_sensitive(False)
+
         self.btbx_lower_buttons.add(self.spn_videodevice)
 
         store = gtk.ListStore(str)
@@ -279,6 +214,7 @@ class CamView(gtk.VBox):
 
     def thread_gtk(self):
         # without this threading function camera speed was realy poor
+
         self.condition = threading.Condition()
         self.thrd = threading.Thread(target=self.run, name="CamView thread")
         self.thrd.daemon = True
@@ -623,70 +559,6 @@ class CamView(gtk.VBox):
             pass
 
 
-class CamProperties():
-    def __init__(self):
-        # get all availible devices
-        self.devices = []
-        self.resolutions = []
-
-    def get_devices(self):
-        result = subprocess.Popen(['v4l2-ctl', '--list-devices'], stdout=subprocess.PIPE).communicate()[0]
-        result = str(result)
-        infos = result.split('\n')
-        for item, info in enumerate(infos):
-            if info == "":
-                continue
-            info = info.strip(' \t\n\r')
-            if "/dev/video" in info:
-                device = str(info + " = " + infos[item - 1])
-                self.devices.append(device)
-        return self.devices
-
-    #    def get_resolution(self, videodevice=0):
-    #        self.resolutions = []
-    #
-    #        # get all availible resolutions
-    #        result = subprocess.Popen(['v4l2-ctl', '-d' + str(videodevice), '--list-formats-ext'], stdout=subprocess.PIPE)#.communicate()[0]
-    #        line = ""
-    #        res =""
-    #        rate = ""
-    #        for letter in result.stdout.read():
-    #            if letter == "\n":
-    #                #print line.strip(" \t\n\r")
-    #                info = line.split(":")
-    #                if "Size" in info[0]:
-    #                    res = info[1].split()
-    #                    #width,height = res[1].split("x")
-    ##                if "Interval" in info[0]:
-    ##                    rate = info[1].split("(")
-    ##                    rate = rate[1]
-    ##                    rate = rate.strip(" fps)")
-    #                    self.resolutions.append(str(res[1]))
-    #                line = ""
-    #            else:
-    #                line += letter
-    #
-    ##        # check for doble entries, not order preserving
-    #        checked = []
-    #        for element in self.resolutions:
-    #            if element not in checked:
-    #                checked.append(element)
-    #
-    #        self.resolutions = sorted(checked)
-    #        return self.resolutions
-
-    def set_powerline_frequeny(self, videodevice, value):
-        command = 'v4l2-ctl -d' + str(videodevice) + ' --set-ctrl=power_line_frequency=' + str(value)
-        self._run_command(command)
-
-    def open_settings(self, videodevice=0):
-        command = "v4l2ucp /dev/video" + str(videodevice)
-        self._run_command(command)
-
-    def _run_command(self, command):
-        if command:
-            result = subprocess.Popen(command, stderr=None, shell=True)
-
 
 # def close_child(self):
 #        os.kill(self.v4l2ucp.pid, signal.SIGKILL)
@@ -695,7 +567,8 @@ class CamProperties():
 def main():
     window = gtk.Window(gtk.WINDOW_TOPLEVEL)
     window.set_title("CamView Window")
-    camv = CamView(videodevice=0, frame_width=640, frame_height=480)
+    video_device = VideoDev(videodevice=0, frame_width=640, frame_height=480)
+    camv = CamView(video_device)
     window.add(camv)
     camv.set_property("draw_color", gtk.gdk.Color("yellow"))
     camv.set_property("circle_size", 150)
